+++ extension.js
@@ -1,36 +1,40 @@
 const vscode                 = require('vscode');
 const CreateComponentCommand = require('./source/ComponentManager');
 const buttonActions          = require('./source/ButtonActions');
-const CreateProjectCommand   = require('./source/ProjectManager');
+const ProjectManager         = require('./source/ProjectManager');
 const ToolsManager           = require('./source/ToolsManager');
+const Logger                 = require('./source/Logger');
+const RefreshConfigsCommand  = require('./source/ConfigManager');
 
 const BuildState      = buttonActions.BuildState;
 const BuildTypes      = buttonActions.BuildTypes;
 const BuildSubsystems = buttonActions.BuildSubsystems;
 
-let buildState = new BuildState(BuildTypes.RELEASE,  BuildSubsystems.NINJA);
+let buildState = new BuildState(BuildTypes.DEBUG,  BuildSubsystems.NINJA);
 
 /**
- * @param {*} context 
+ * @param {vscode.ExtensionContext} context The extension context provided by VSCode.
  */
 function activate(context) 
 {
     const buttons = 
     [
-        new StatusBarButton("Build Type", `$(gear) ${buildState.type}`, "extension.selectBuild", "Click to switch build type",  7),
-        new StatusBarButton("Clean",      "$(trash) Clean",             "extension.clean",       "Clean the build",             6),
-        new StatusBarButton("Build",      "$(database) Build",          "extension.build",       "Build the project",           5),
-        new StatusBarButton("Run"  ,      "$(run) Run",                 "extension.run",         "Run the application",         4),
-        new StatusBarButton("Debug",      "$(debug) Debug",             "extension.debug",       "Debug the processor",         3),
-        new StatusBarButton("Test",       "$(beaker) Test",             "extension.test",        "Run tests",                   2),
-        new StatusBarButton("Debug Test", "$(debug-alt) Debug Test",    "extension.debugTest",   "Click to debug the test app", 1)
+        new StatusBarButton("Build Type", `$(gear) ${buildState.type}`, "extension.selectBuild", "Click to switch build type",  16),
+        new StatusBarButton("Clean",      "$(trash) Clean",             "extension.clean",       "Clean the build",             15),
+        new StatusBarButton("Build",      "$(database) Build",          "extension.build",       "Build the project",           14),
+        new StatusBarButton("Run"  ,      "$(run) Run",                 "extension.run",         "Run the application",         13),
+        new StatusBarButton("Debug",      "$(debug) Debug",             "extension.debug",       "Debug the processor",         12),
+        new StatusBarButton("Test",       "$(beaker) Test",             "extension.test",        "Run tests",                   11),
+        new StatusBarButton("Debug Test", "$(debug-alt) Debug Test",    "extension.debugTest",   "Click to debug the test app", 10)
     ];
 
     const disposables = buttons.map(button => createStatusBarItem(button, context));
 
     CreateComponentCommand(context);
-    CreateProjectCommand(context);
-    ToolsManager.searchForTools();
+    ProjectManager.CreateProjectCommand(context);
+    RefreshConfigsCommand(context);
+    ToolsManager.SearchForToolsCommand(context);
+    ToolsManager.SearchForTools();
 
     vscode.window.onDidChangeActiveColorTheme(e => // eslint-disable-line no-unused-vars
     {
@@ -64,7 +68,7 @@ class StatusBarButton
 
 /**
  * @param {StatusBarButton} button
- * @param {*} context
+ * @param {vscode.ExtensionContext} context The extension context provided by VSCode.
  * @returns {vscode.StatusBarItem}
  */
 function createStatusBarItem(button, context)
@@ -104,7 +108,7 @@ function createStatusBarItem(button, context)
 
 function deactivate()
 {
-    console.log("C Toolkit extension deactivated");
+    Logger.Info("C Toolkit extension deactivated");
 }
 
 module.exports =