+++ source/ToolsManager.js
@@ -1,7 +1,27 @@
 const vscode = require('vscode');
 const os     = require('os');
-const utils  = require('./Utils');
-const { exec, spawn, spawnSync } = require('child_process');
+const Logger = require('./Logger');
+const { execSync } = require('child_process');
+const { spawn, spawnSync } = require('child_process');
+const { OsTypes, CheckOs, FormatList, WrapSpacedComponents } = require('./CommonUtils');
+
+/** @type {vscode.Disposable} */
+let searchForToolsDisposable;
+
+/**
+ * Registers the 'searchForTools' command in the extension.
+ * 
+ * @param {vscode.ExtensionContext} context The extension context provided by VSCode.
+ */
+function SearchForToolsCommand(context)
+{
+    if (searchForToolsDisposable)
+    {
+        searchForToolsDisposable.dispose();
+    }
+    searchForToolsDisposable = vscode.commands.registerCommand("extension.searchForTools", SearchForTools);
+    context.subscriptions.push(searchForToolsDisposable);
+}
 
 /**
  * Enum for build tools required
@@ -13,14 +33,19 @@ const BuildTools =
     CMAKE : 'cmake',
     NINJA : 'ninja',
     MAKE  : 'make',
-    SCOOP : 'scoop'
+    SCOOP : 'scoop',
+    GIT   : 'git',
+    BBOX  : 'busybox',
 };
 
+/**
+ * Enum for package managers
+ */
 const PackageManagers = 
 {
-    [utils.OsTypes.WINDOWS] : 'scoop',
-    [utils.OsTypes.LINUX]   : 'apt-get',
-    [utils.OsTypes.MACOS]   : 'brew',
+    [OsTypes.WINDOWS] : 'scoop',
+    [OsTypes.LINUX]   : 'apt-get',
+    [OsTypes.MACOS]   : 'brew',
 };
 
 /**
@@ -32,22 +57,21 @@ const PackageManagers =
  */
 async function isToolInPath(toolName)
 {
-    toolName = toolName === BuildTools.SCOOP ? `${os.homedir()}\\scoop\\shims\\scoop` : toolName;
-    return new Promise((resolve, reject) => 
+    toolName = toolName === BuildTools.SCOOP ? `${WrapSpacedComponents(os.homedir())}\\scoop\\shims\\scoop` : toolName;
+
+    // Busybox does not support '--version' argument so this case ahs been added for it
+    const versionCommand = toolName === BuildTools.BBOX ? toolName : `${toolName} --version`;
+
+    try
     {
-        exec(`${toolName} --version`, (error, stdout, stderr) => 
-        {
-            if (error) 
-            {
-                resolve(false);
-            }
-            else
-            {
-                resolve(true);
-            }
-            console.log(`error: ${error}, stdout: ${stdout}, stderr: ${stderr}`);
-        });
-    });
+        execSync(versionCommand, { stdio: 'ignore' });
+        Logger.Info(`${toolName} found.`);
+        return true;
+    }
+    catch (error)
+    {
+        return false;
+    }
 }
 
 /**
@@ -58,6 +82,8 @@ async function isToolInPath(toolName)
  */
 const installScoop = () => 
 {
+    Logger.Info('Attempting to install Scoop');
+
     const installCommand = 'powershell';
     const installArgs = ['-Command', '& {Set-ExecutionPolicy RemoteSigned -scope CurrentUser; iwr -useb get.scoop.sh | iex}'];
 
@@ -67,12 +93,12 @@ const installScoop = () =>
 
         process.stdout.on('data', (data) => 
         {
-            console.log(`stdout: ${data}`);
+            Logger.Info(`stdout: ${data}`);
         });
 
         process.stderr.on('data', (data) => 
         {
-            console.error(`stderr: ${data}`);
+            Logger.Error(`stderr: ${data}`);
         });
 
         process.on('close', (code) => 
@@ -104,7 +130,7 @@ async function askAndInstallScoop()
                 location: vscode.ProgressLocation.Notification,
                 title: `Installing ${toolName}`,
                 cancellable: false
-            }, async (progress, token) =>
+            }, async (progress, token) => // eslint-disable-line no-unused-vars
             {
                 progress.report({ message: `In Progress...` });
                 try
@@ -123,6 +149,7 @@ async function askAndInstallScoop()
         }
         else if (selection === 'No')
         {
+            Logger.Warning('Scoop was not found and the user did not consent to installation');
             vscode.window.showWarningMessage('Build tools for Windows would have to be installed manually.');
         }
     });
@@ -131,21 +158,23 @@ async function askAndInstallScoop()
 /**
  * Executes a specified command with provided arguments.
  * 
- * @param {string} command The command to execute.
- * @param {string[]} args The arguments for the command.
+ * @param {string?}  userPassword Password provided by the user for tool installation.
+ * @param {string}   command      The command to execute.
+ * @param {string[]} args         The arguments for the command.
  * 
  * @returns {string} Stdout string if it passes.
  */
-function execCommand(command, args)
+function execCommand(userPassword, command, args)
 {
     if (command === 'scoop')
     {
         command = 'cmd';
-        args = ['/c', 'scoop', ...args];
+        args = ['/c', `${WrapSpacedComponents(os.homedir())}\\scoop\\shims\\scoop`, ...args];
     }
 
-    console.log(`Executing: ${command} ${args.join(' ')}`);
-    const process = spawnSync(command, args);
+    Logger.Info(`Executing: ${command} ${args.join(' ')}`);
+
+    const process = spawnSync(command, args, { input: Buffer.from(`${userPassword}\n`, "utf-8"), });
 
     // Convert Buffer to String
     const stdout = process.stdout ? process.stdout.toString() : '';
@@ -154,12 +183,12 @@ function execCommand(command, args)
     // Log the outputs
     if (stdout)
     {
-        console.log(`stdout: ${stdout}`);
+        Logger.Info(`stdout: ${stdout}`);
     }
 
     if (stderr)
     {
-        console.error(`stderr: ${stderr}`);
+        Logger.Error(`stderr: ${stderr}`);
     }
 
     if (process.status !== 0)
@@ -173,18 +202,21 @@ function execCommand(command, args)
 /**
  * Initiates the installation of a specified tool by using the appropriate tool manager.
  * 
- * @param {string} toolName The name of the tool to install.
+ * @param {string}  toolName     The name of the tool to install.
+ * @param {string?} userPassword Password provided by the user for tool installation.
  * 
  * @returns {Thenable<boolean>} I don't know what this is. I miss C where bool is literally an int.
  */
-function installTool(toolName)
+function installTool(toolName, userPassword)
 {
     return vscode.window.withProgress({
         location: vscode.ProgressLocation.Notification,
         title: `Installing ${toolName}`,
         cancellable: false
-    }, async (progress, token) =>
+    }, async (progress, token) => // eslint-disable-line no-unused-vars
     {
+        Logger.Info(`Attempting to install ${toolName}`);
+
         let isInstalled = false;
         progress.report({ message: `In Progress...` });
 
@@ -193,7 +225,7 @@ function installTool(toolName)
             // Scoop installation has its own function.
             if (toolName === BuildTools.SCOOP) throw new Error("Scoop is not supposed to be installed from installTool() function");
 
-            let installCommand = PackageManagers[utils.CheckOs()] || null;
+            let installCommand = PackageManagers[CheckOs()] || null;
             if (installCommand === null) throw new Error('Unsupported OS type');
 
             /**
@@ -203,9 +235,22 @@ function installTool(toolName)
             const installArgs = ['install', toolName];
 
             // this has not been tested yet.
-            if (installCommand === 'apt-get') installArgs.unshift('sudo');
+            if (installCommand === 'apt-get')
+            {
+                // sanity check
+                if (userPassword === undefined) throw new Error('No password was provided to installTool while using apt.');
 
-            execCommand(installCommand, installArgs);
+                installCommand = 'sudo';
+                installArgs.unshift('apt-get');
+                installArgs.unshift('-S');
+                installArgs.push('-y');
+
+                // in APT, 'ninja' is 'ninja-build'
+                const indexOfNinja = installArgs.indexOf('ninja');
+                if (indexOfNinja !== -1) installArgs[indexOfNinja] = 'ninja-build';
+            }
+
+            execCommand(userPassword, installCommand, installArgs);
             isInstalled = true;
             progress.report({ message: `Finalizing...` });
         }
@@ -243,9 +288,10 @@ function askForRestart(restartReason)
  * it asks the user if they want to restart VSCode. If allowed, it restarts VSCode. 
  * If any tools fail to install, it notifies the user about those tools.
  * 
- * @param {string[]} tools An array containing names of the tools to install.
+ * @param {string[]} tools        An array containing names of the tools to install.
+ * @param {string?}  userPassword Password provided by the user for tool installation.
  */
-async function InstallMultipleTools(tools)
+async function InstallMultipleTools(tools, userPassword)
 {
     /** @type {string[]} */
     let installedTools = [];
@@ -257,12 +303,7 @@ async function InstallMultipleTools(tools)
     {
         try
         {
-            if ((tool === BuildTools.GDB) && (utils.CheckOs() === utils.OsTypes.MACOS))
-            {
-                throw new Error('GDB will not be used for MacOS, instead LLDB support will be added.');
-            }
-
-            let selection = await installTool(tool);
+            let selection = await installTool(tool, userPassword);
             if (selection === true)
             {
                 installedTools.push(tool);
@@ -286,8 +327,11 @@ async function InstallMultipleTools(tools)
  * For now, this strange logic handles the failed installation messages thrown by multiple failed
  * installations, while also showing the user why the individual installation failed.
  * 
- * @param {string[]} installedTools
- * @param {Array<{tool: string, failReason: string}>} failedTools
+ * @param {string[]} installedTools 
+ * Array containing names of the tools for which installations passed.
+ * 
+ * @param {Array<{tool: string, failReason: string}>} failedTools 
+ * Array containing names and fail reasons of tools for which installations failed.
  */
 function processInstallationOutputs(installedTools, failedTools)
 {
@@ -300,15 +344,39 @@ function processInstallationOutputs(installedTools, failedTools)
             let errorStr = `${eachError.tool} (${eachError.failReason})`;
             failArray.push(errorStr);
         }
-        const failedToolList = utils.FormatList(failArray);
+        const failedToolList = FormatList(failArray);
         vscode.window.showErrorMessage(`Failed to install: ${failedToolList}. Please install manually.`);
     }
 
     if (installedTools.length > 0)
     {
-        const installedToolList = utils.FormatList(installedTools);
+        const installedToolList = FormatList(installedTools);
         askForRestart(`${installedToolList} installation completed`);
     }
+}
+
+/**
+ * Asks the user their sudo password to allow for tools installation.
+ * 
+ * @returns {Promise<string>} User's password
+ */
+async function askForPassword()
+{
+    // Ask the user about the project name
+    let passwordAsker = await vscode.window.showInputBox({
+        prompt     : 'Please enter your sudo password, it is required to to install the tools.',
+        value      : undefined,
+        placeHolder: 'your password',
+        password   : true,
+    });
+
+    if (!passwordAsker)
+    {
+        vscode.window.showErrorMessage('Superuser access is required to install the missing tools.');
+        return;
+    }
+
+    return passwordAsker;
 }
 
 /**
@@ -319,30 +387,37 @@ function processInstallationOutputs(installedTools, failedTools)
  */
 async function askAndInstallMultipleTools(tools)
 {
-    const toolList = utils.FormatList(tools);
-    vscode.window.showWarningMessage(`${toolList} not found. Would you like to install?`, 'Yes', 'No').then(async selection =>
+    const toolList = FormatList(tools);
+
+    const selection = await vscode.window.showWarningMessage(`${toolList} not found. Would you like to install?`, 'Yes', 'No');
+    if (selection === 'Yes')
     {
-        if (selection === 'Yes')
+        /** @type {string} */
+        let userPassword = undefined;
+        if (CheckOs() === OsTypes.LINUX)
         {
-            InstallMultipleTools(tools);
+            userPassword = await askForPassword();
         }
-        else if (selection === 'No')
-        {
-            vscode.window.showWarningMessage(`${toolList} would have to be installed manually.`);
-        }
-    });
+        await InstallMultipleTools(tools, userPassword);
+    }
+    else
+    {
+        Logger.Warning(`${toolList} not found and the user did not consent to installation`);
+        vscode.window.showWarningMessage(`${toolList} would have to be installed manually.`);
+    }
 }
 
 /**
+ * Searches for build tools and on Windows also for the package manager 'Scoop'.
  * Prompts the user for installing Scoop. If the user agrees, initiates the installation.
  * After installation, it either displays a success message or an error message based on the outcome.
  */
-async function searchForTools()
+async function SearchForTools()
 {
     /** @type {string[]} */
     let missingTools = [];
 
-    if (utils.CheckOs() === utils.OsTypes.WINDOWS)
+    if (CheckOs() === OsTypes.WINDOWS)
     {
         let scoop = await isToolInPath(BuildTools.SCOOP);
         if (scoop === false)
@@ -352,32 +427,28 @@ async function searchForTools()
         }
     }
 
-    let results = await Promise.all
-    ([
-        isToolInPath(BuildTools.GCC),
-        isToolInPath(BuildTools.GDB),
-        isToolInPath(BuildTools.CMAKE),
-        isToolInPath(BuildTools.NINJA),
-        isToolInPath(BuildTools.MAKE)
-    ]);
+    let toolsToCheck = [BuildTools.GCC, BuildTools.CMAKE, BuildTools.NINJA, BuildTools.MAKE, BuildTools.GIT];
+    if (CheckOs() !== OsTypes.MACOS) toolsToCheck.push(BuildTools.GDB);
+    if (CheckOs() === OsTypes.WINDOWS) toolsToCheck.push(BuildTools.BBOX);
 
-    const tools = [BuildTools.GCC, BuildTools.GDB, BuildTools.CMAKE, BuildTools.NINJA, BuildTools.MAKE];
+    let results = await Promise.all(toolsToCheck.map(tool => isToolInPath(tool)));
 
     for (let i = 0; i < results.length; i++)
     {
         if (!results[i])
         {
-            missingTools.push(tools[i]);
+            missingTools.push(toolsToCheck[i]);
         }
     }
 
     if (missingTools.length > 0)
     {
-        askAndInstallMultipleTools(missingTools);
+        await askAndInstallMultipleTools(missingTools);
     }
 }
 
 module.exports = 
 {
-    searchForTools,
+    SearchForToolsCommand,
+    SearchForTools
 };